Turn Based System Functional (Hard-Coded until other classes are set up)
Currently got a WIP Turn based system. The Enemy intent and attack phase will be improved on when the enemy class is created and worked on so i gain references there related to logic of moving and attacking. THe player action also follows the same suit. Functionality is hardcoded to see process through the phases.